refactor: merge Strategy domain into Blueprints
The Strategy domain held a single page -- the blueprints boards overview at
/strategy/showBoards -- whose service already delegated to the Blueprints
service. There is no other use for a separate "strategy" module in core, so
fold it in and delete the domain.

Changes:
- New native controller Blueprints\Controllers\ShowBoards (plain class,
  constructor DI, get(): Response) matching the converted Blueprints
  controllers; renders blueprints.showBoards.
- Strategy\Services\Strategy's 4 methods folded into Blueprints\Services\
  Blueprints (getStrategyBoardsOverview -> getBoardsOverview); the former
  cross-domain calls become self-calls (getLastUpdatedCanvas/getBoardProgress).
- Template moved to Blueprints/Templates/showBoards.blade.php (no body edits).
- Route Route::get('/blueprints/showBoards', [ShowBoards, 'get']) plus a 301
  redirect from the old /strategy/showBoards (bookmarks + not-yet-updated
  plugins).
- Rewired all core references: menu item, help modal key, project checklist
  link, and the three DelCanvas redirects (Blueprints/Canvas/Goalcanvas).
- Deleted app/Domain/Strategy entirely.

Menu: the left-nav Blueprints item now uses module 'blueprints' with no
'active' key, so it highlights on EVERY /blueprints/* URL (the wiki/files
idiom) -- previously it only lit on /strategy/showBoards. The default-menu
Retros quick-link was dropped (it sat in the same submenu and would have
double-highlighted); retros remains in the full menu and reachable from the
boards overview.

Routing + DI verified via the boot harness; phpstan + pint clean. The
StrategyPro/PgmPro plugin references to /strategy/showBoards live in the
app/Plugins submodule and must be updated there separately.
